refactor: add index entry
