+++ src/library/thingsboardConnecter/device/index.ts
@@ -1,0 +1,3 @@
+export { default as createTBDevice } from './createDevice';
+export { default as deleteTBDevice } from './deleteDevice';
+export { default as getTBDevice } from './getDevices';
